fix(fill): truncate infill offsets at the source coord_t boundary
ExPolygonWithOffset takes coord_t aoffset1/aoffset2 (FillRectilinear
.cpp:391-397), so the float(scale_(...)) arguments round to f32 and then
truncate to integer microns before ClipperOffset runs. The port kept
the fractional f32 deltas end to end, shifting inner-shrink vertices by
a fraction that flips 1um rounding boundaries.

Chain the conversion exactly like the source: f64 -> f32 -> coord_t ->
f32. With the truncation restored the fill shrink deltas match the
oracle bit-for-bit (a1=-22500, a2=-225000, diff=202500 exact), and the
G-code body's first divergence moves from line 1853 to 3624.

+++ crates/ares-core/src/fill/rectilinear/surface.rs
@@ -105,12 +105,14 @@ pub(super) fn scaled_offsets(
     ))
 }
 
+// ExPolygonWithOffset receives coord_t parameters: the float offsets truncate to
+// integer microns before ClipperOffset sees them (FillRectilinear.cpp:391-397).
 fn checked_scaled_f32(scale: CoordinateScale, value: f64) -> Result<f32, ClipperError> {
     let scaled = value / scale.factor();
     if !scaled.is_finite() || !(i64::MIN as f64..-(i64::MIN as f64)).contains(&scaled) {
         return Err(ClipperError::CoordinateOutOfRange);
     }
-    Ok(scaled as f32)
+    Ok(scaled as f32 as i64 as f32)
 }
 
 fn adjust_solid_spacing(width: i64, distance: i64) -> i64 {
